+++ src/Firebase/Exception/AppCheckException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kreait\Firebase\Exception;
+
+interface AppCheckException extends FirebaseException
+{
+}